Manage correctly "legendary" executable
A system wide Python 3 is required. If "Legendary" is present as a Python Script, the wrapper exit, because the Python Script version cannot run inside Steam Linux Runtime (missing Python libraries), and it's not possible to safely mix the Steam Linux Runtime Python libraries, with the host Python libraries (presents unser /run/host/usr...).
If is present a binary version of "Legendary" the wrapper set correct PYTHONHOME and PYTHONPATH inside the Steam Linux Runtime, and set the correct path to reach the binary executable inside the sandbox (/run/host/usr... if the binary is under /usr, otherwise it leave it untouched and set PRESSURE_VESSEL_FILESYSTEMS_RO varibale to bind the binary inside the sandbox).